divides the diary
schemes and controllers in separate files
