+++ models/diary/index.js
@@ -1,0 +1,7 @@
+const meal = require('./meal');
+const workout = require('./workout');
+
+module.exports = {
+	meal,
+  workout,
+};
