+++ models/index.js
@@ -1,4 +1,3 @@
-// пример реэкспорта
 const diary = require('./diary');
 // const users = require('./users');
