+++ routes/api/diary.js
@@ -1,6 +1,6 @@
 const express = require('express');
 const { ctrlDiary } = require('../../controllers');
-const { auth } = require('../../middlewares');
+const { auth, isValidId } = require('../../middlewares');
 
 const router = express.Router();
 
@@ -8,7 +8,7 @@ router.get('/meal', auth, ctrlDiary.getByDateMeal);
 router.get('/workout', auth, ctrlDiary.getByDateWorkout);
 router.post('/meal', auth, ctrlDiary.postAddMeal);
 router.post('/workout', auth, ctrlDiary.postAddWorkout);
-router.delete('/meal', auth, ctrlDiary.deleteByNameAndDateMeal);
-router.delete('/workout', auth, ctrlDiary.deleteByNameAndDateWorkout);
+router.delete('/meal/:id', auth, isValidId, ctrlDiary.deleteByIdMeal);
+router.delete('/workout/:id', auth, isValidId, ctrlDiary.deleteByIdWorkout);
 
 module.exports = router
